+++ ProcessFileEdit.js
@@ -1,18 +1,24 @@
-$(document).ready( function() {
+$(document).ready(function() {
 
 	// Hide all subfolders at startup
-	//MP $(".php-file-tree").find("ul").hide(); I hide them in css, much faster
+	// $(".php-file-tree").find("ul").hide(); I hide them with css, much faster
 
 	// Expand/collapse on click
-	$(".pft-d a").click( function() {
-		$(this).parent().find("ul:first").slideToggle("fast");
-		if( $(this).parent().attr('class') == "pft-d" ) {
+	$(".pft-d a").click(function(e) {
+		var parent = $(this).parent();
+		e.preventDefault();
+		parent.find("ul:first").slideToggle("fast");
+		parent.toggleClass("pft-d-open");
+		//if(parent().attr('class') == "pft-d") {
+		if(parent.hasClass("pft-d")) {
 			return false;
 		};
 	});
 
 	function setupCloneButton1() {
-		//MP this is a copy of main.js from admin theme
+		// this is a copy of main.js from admin theme
+
+		if($("body").is(".modal") == false) return; // prevent doubling in main.js
 
 		// if there are buttons in the format "a button" without ID attributes, copy them into the masthead
 		// or buttons in the format button.head_button_clone with an ID attribute.
@@ -20,8 +26,7 @@ $(document).ready( function() {
 		// var $buttons = $("#content a:not([id]) button:not([id]), #content button.head_button_clone[id!=]");
 		var $buttons = $("button.head_button_clone, button.head-button");
 
-		// don't continue if no buttons here or if we're in IE
-		if($buttons.length == 0) return; // || $.browser.msie) return;
+		if($buttons.length == 0) return;
 
 		var $head = $("#head_button");
 		//MP if($head.length == 0) $head = $("<div id='head_button'></div>").prependTo("#breadcrumbs .container");
@@ -50,7 +55,10 @@ $(document).ready( function() {
 
 	setupCloneButton1();
 
-	saveButton = $("#saveFile");
+	var saveButton = $("#saveFile"),
+			isChanged = $("#change"),
+			closeBtn = parent.$(".ui-dialog-titlebar-close");
+
 	saveButton.on('click', function (e) {
 		e.preventDefault();
 		window.editor.save();
@@ -61,7 +69,9 @@ $(document).ready( function() {
 			type: "POST",
 			success: function (data) {
 				if(data === "") {
-					$("#change").html(""); //remove changes indicator (*)
+					isChanged.html(""); //remove changes indicator (*)
+					$('#saveFile').addClass('ui-state-disabled');
+					$('#saveFile_copy').addClass('ui-state-disabled');
 				} else {
 					alert(data);
 				}
@@ -74,6 +84,29 @@ $(document).ready( function() {
 		});
 
 		return false;
+	});
+
+	$(document).on('keydown', function (e) {
+		e = e || window.event;
+		if(e.keyCode === 27 && closeBtn.length) {
+			//e.stopImmediatePropagation();
+			closeBtn.trigger('mousedown');
+		}
+	});
+
+	closeBtn.on('mousedown', function (e) {
+		//if(isEditFormChanged) {
+		if(isChanged.html() !== "") {
+			var confirm = window.confirm('File is not saved. Continue?');
+			if(!confirm) {
+				e.preventDefault();
+				return false;
+			}
+			//isEditFormChanged = false;
+			isChanged.html("");
+			//$(this).click(); //not ok
+			closeBtn.click();
+		}
 	});
 
 });